JTAND-6910 Maven publish plugin needs to have the repo defined explicitly
It does not take the configuration from the init.gradle script. After the discussion with delta, as the publishing is done using the Jenkins job, the credentials are not required (from local address). Also it is recommended (in order to be consistent  with backend libs) to hardcode the artifactory url into the project rather than define it on the machine level (init.gradle way).